refactor: Product 페이지 suspense 적용
