+++ src/routes/Routers.tsx
@@ -8,6 +8,8 @@ import HOME from '@/page/Home';
 import ThemesPage from '@/page/Themes';
 import { ROUTE_PATH } from './routePath';
 import ProductPage from '@/page/Product';
+import { Suspense } from 'react';
+import Loading from '@/components/Loading';
 
 const AppRouter = () => {
   return (
@@ -26,7 +28,9 @@ const AppRouter = () => {
         path={ROUTE_PATH.ORDER}
         element={
           <RequireAuth>
-            <OrderPage />
+            <Suspense fallback={<Loading />}>
+              <OrderPage />
+            </Suspense>
           </RequireAuth>
         }
       />
